Made a simple nav bar with a svg Logo
So imporatant thing to read:
Since we are currently not using a real frame work like react.js, angular.js, or vue.js...everytime we make an edit to the nav bar we have to make the same changes everywhere else in the other pages. Frameworks like react and vue aim to solve this issue but the learning curve can be a bit more difficult since its a lot more java script than easy html stuff. that is all.